Splitting kernel_iteration_layout
